pyproject.toml
[project]
name = "feedback-report-generator"
version = "0.1.0"
description = "Feedback Report Generator"
authors = [
    {name = "Azhar Bihari"}
]
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "celery[redis] (>=5.5.2,<6.0.0)",
    "markdown (>=3.8,<4.0)",
    "reportlab (>=4.4.0,<5.0.0)",
    "django-filter (>=25.1,<26.0)",
    "flower (>=2.0.1,<3.0.0)",
    "psycopg[binary] (>=3.2.7,<4.0.0)",
    "pydantic (>=2.11.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false